Security

+++ src/app/add-conge/add-conge.component.ts
@@ -6,6 +6,7 @@ import { Conge } from '../model/conge.model';
 import { CongeService } from '../services/conge.service';
 import { User } from '../model/user.model';
 import { NgbDate } from '@ng-bootstrap/ng-bootstrap';
+import { AuthenticationService } from '../services/authentication.service';
 
 @Component({
   selector: 'app-add-conge',
@@ -14,12 +15,11 @@ import { NgbDate } from '@ng-bootstrap/ng-bootstrap';
 })
 export class AddCongeComponent implements OnInit {
 
-  etat: boolean=false;
   conge: Conge ; 
   edit_mode: string;
   id: number;
-  user = new User(1,'user','user','Baha','Zaghdoudi','zagdoudi@gmail.com','98654321',true,null,null,null,null);
-  editValue = new Conge(null,null,null,'',false,false,0,null);
+  userAuthenticated = new User(-1,'','','','','','',false,null,null,null,null,0);
+  editValue = new Conge(null,null,null,'',false,false,null);
 
 
   @ViewChild('formulaire') formulaire : NgForm;
@@ -28,13 +28,15 @@ export class AddCongeComponent implements OnInit {
               private router: Router,
               private route : ActivatedRoute,
               private dialog : DialogService,
+              public authService : AuthenticationService,
                ) { }
 
   ngOnInit(): void {
+    this.authService.getLogin();
+    this.userAuthenticated = this.authService.getUserAuthenticated();
 
     this.route.queryParams.subscribe(qp =>{
-      this.edit_mode=qp['edit_mode'];
-      console.log(this.editValue);
+      this.edit_mode = qp['edit_mode'];
     });
 
     if(this.edit_mode == "modifier"){
@@ -47,51 +49,39 @@ export class AddCongeComponent implements OnInit {
         },
         error => {
           console.log(error);
-          
         }
-      )
+      );
     }
   }
 
   saveConge(){
     this.conge = this.formulaire.form.value;  
-    console.log(this.conge)  
+  
     /* Covert DatePicker to Date */
     const datedeb = this.toDate(this.formulaire.form.get('dateDebut').value);
     const dateF = this.toDate(this.formulaire.form.get('dateFin').value)
     this.conge.dateDebut = datedeb;
     this.conge.dateFin = dateF;
-
-    this.conge.confirmation = this.etat;
-    this.conge.user = this.user;
-
+    this.conge.user = this.userAuthenticated;
+    console.log(this.userAuthenticated);
+    console.log(this.conge.user);
+    
+    
+    
      if(this.edit_mode == 'modifier'){
-        this.conge.id = this.id;
-        this.congeService.getCongeById(this.id).subscribe (
-          conge => {
-            this.editValue = conge
-            console.log(this.editValue); 
-          },
-          error => {
-            console.log(error);
-          }
-        );
-        this.conge = this.editValue;
-
+        this.conge.id = this.id; 
       }
+
+      
+
       this.congeService.saveConge(this.conge).subscribe (
         response => {
-          const link = ['home/conge']
-          this.router.navigate (link);
-          console.log(response);
+          this.router.navigate (['home/conge']);
         },
         error=> {
          console.log(error);
         }
-      )
-   }
-  changeStatus(status){
-    this.etat = status;
+        );
   }
 
   onClear() {
@@ -99,17 +89,15 @@ export class AddCongeComponent implements OnInit {
   }
 
   onDelete(){
-
-      this.dialog.openConfirmDialog('Vous êtes sure de supprimer ce congé ?')
+    this.dialog.openConfirmDialog('Vous êtes sure de supprimer ce congé ?')
       .afterClosed().subscribe(res =>{
         if(res){
           this.congeService.deleteConge(this.id).subscribe(
             response =>{
             this.router.navigate (['home/conge']);
-            console.log(response);
           });
         }
-      });
+    });
   }
  
   toDate(date : NgbDate): Date{
@@ -117,4 +105,15 @@ export class AddCongeComponent implements OnInit {
                     date.month -1, 
                     date.day +1);
   }
+
+  accepter(){
+    console.log(this.editValue);
+    
+    
+    this.congeService.acceptConge(this.editValue,this.editValue.user.login).subscribe(resp =>{
+      this.router.navigate (['home/conge']);
+    },err =>{console.log(err);
+    })
+  }
+
 }
